+++ src/core/mod.rs
@@ -3,20 +3,11 @@ use bevy::prelude::*;
 pub mod effects;
 pub mod resources;
 
-use self::{
-    effects::{animate::AnimatePlugin, blink::EffectsPlugin},
-    resources::{despawn_timer::DespawnTimerPlugin, game_time::GameTimePlugin, state::StatePlugin},
-};
+use self::resources::game_time::GameTimePlugin;
 
 pub struct CorePlugin;
 impl Plugin for CorePlugin {
     fn build(&self, app: &mut App) {
-        app.add_plugins((
-            GameTimePlugin,
-            StatePlugin,
-            EffectsPlugin,
-            AnimatePlugin,
-            DespawnTimerPlugin,
-        ));
+        app.add_plugins(GameTimePlugin);
     }
 }