+++ src/core/resources/mod.rs
@@ -2,4 +2,3 @@ pub mod assets;
 pub mod config;
 pub mod despawn_timer;
 pub mod game_time;
-pub mod state;