+++ src/ui/menus/pause.rs
@@ -3,33 +3,24 @@ use bevy_rapier2d::prelude::RapierConfiguration;
 use leafwing_input_manager::prelude::{ActionState, InputManagerPlugin};
 
 use crate::{
-    core::{
-        effects::blink::DrawBlinkTimer,
-        resources::{
-            assets::UiAssets,
-            state::{is_in_game_state, ForState, GameState},
-        },
-    },
+    core::{effects::blink::DrawBlinkTimer, resources::assets::UiAssets},
     inputs::pause::{pause_input_map, PauseAction},
+    systems::states::{ForState, GameState},
 };
 
 pub struct PausePlugin;
 impl Plugin for PausePlugin {
     fn build(&self, app: &mut App) {
         app.add_plugins(InputManagerPlugin::<PauseAction>::default());
-
-        app.add_systems(OnEnter(GameState::GameCreate), setup)
-            .add_systems(Update, pause_system.run_if(is_in_game_state))
-            .add_systems(OnEnter(GameState::Paused), pause_screen);
     }
 }
 
-fn setup(mut commands: Commands) {
+pub(crate) fn setup_pause_systems(mut commands: Commands) {
     commands.insert_resource(pause_input_map());
     commands.insert_resource(ActionState::<PauseAction>::default());
 }
 
-fn pause_system(
+pub(crate) fn pause_system(
     state: Res<State<GameState>>,
     mut next_state: ResMut<NextState<GameState>>,
     pause_action_state: Res<ActionState<PauseAction>>,
@@ -50,7 +41,7 @@ fn pause_system(
     }
 }
 
-fn pause_screen(mut commands: Commands, ui: Res<UiAssets>) {
+pub(crate) fn pause_screen(mut commands: Commands, ui: Res<UiAssets>) {
     commands
         .spawn((
             NodeBundle {