+++ src/ui/menus/start_menu.rs
@@ -7,31 +7,20 @@ use leafwing_input_manager::prelude::{ActionState, InputManagerPlugin};
 use crate::{
     core::{
         effects::blink::DrawBlinkTimer,
-        resources::{
-            assets::{AudioAssets, UiAssets},
-            state::{is_in_menu_state, ForState, GameState},
-        },
+        resources::assets::{AudioAssets, UiAssets},
     },
     inputs::menu::{menu_input_map, MenuAction},
+    systems::states::{ForState, GameState},
 };
 
 pub struct MenuPlugin;
 impl Plugin for MenuPlugin {
     fn build(&self, app: &mut App) {
         app.add_plugins(InputManagerPlugin::<MenuAction>::default());
-        app.add_systems(
-            OnTransition {
-                from: GameState::Loading,
-                to: GameState::StartMenu,
-            },
-            init,
-        )
-        .add_systems(OnEnter(GameState::StartMenu), setup)
-        .add_systems(Update, menu_input_system.run_if(is_in_menu_state));
     }
 }
 
-fn init(mut commands: Commands, audios: Res<AudioAssets>) {
+pub(crate) fn init_start_menu(mut commands: Commands, audios: Res<AudioAssets>) {
     commands.insert_resource(menu_input_map());
     commands.insert_resource(ActionState::<MenuAction>::default());
 
@@ -51,7 +40,7 @@ fn init(mut commands: Commands, audios: Res<AudioAssets>) {
     ));
 }
 
-fn setup(mut commands: Commands, ui: Res<UiAssets>) {
+pub(crate) fn spawn_start_menu(mut commands: Commands, ui: Res<UiAssets>) {
     commands
         .spawn((
             NodeBundle {
@@ -154,7 +143,7 @@ fn setup(mut commands: Commands, ui: Res<UiAssets>) {
         });
 }
 
-fn menu_input_system(
+pub(crate) fn menu_input_system(
     state: Res<State<GameState>>,
     mut next_state: ResMut<NextState<GameState>>,
     inputs: Res<ActionState<MenuAction>>,