+++ src/ui/mod.rs
@@ -8,21 +8,12 @@ pub mod resources;
 
 use self::{
     camera::CameraPlugin,
-    damage::UiDamagePlugin,
-    hud::HudPlugin,
     menus::{credits::CreditsPlugin, pause::PausePlugin, start_menu::MenuPlugin},
 };
 
 pub struct UiPlugin;
 impl Plugin for UiPlugin {
     fn build(&self, app: &mut App) {
-        app.add_plugins((
-            HudPlugin,
-            MenuPlugin,
-            CreditsPlugin,
-            PausePlugin,
-            CameraPlugin,
-            UiDamagePlugin,
-        ));
+        app.add_plugins((MenuPlugin, CreditsPlugin, PausePlugin, CameraPlugin));
     }
 }