+++ CHANGELOG.md
@@ -7,6 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Migrated core, temp, ui, world and ship systems to systems/mod.rs file for better overview
+- Move set_window_icon temporary system to PreStartup schedule
+- Move configure_physics_engine to GameCreate state
+- Move system sets to new systems module
+- Move state structs and transitions to new systems module
+- Move spawn and contact events to systems::events submodule
+
+### Fixed
+
+- enemy_targeting_system now happens before ship_damage, preventing crash to desktop
+
 ## [0.0.19] - 2023-10-31
 
 ### Added